change tests directory structure
