+++ api/app/tests/requests/upload_test/no_tag_key.jsonl
@@ -1,0 +1,2 @@
+{"t":"teststring"}
+{"tag_name":"alpha"}
