+++ api/app/tests/requests/upload_test/tag.jsonl
@@ -1,0 +1,5 @@
+{"tag_name":"teststring","references":[{"target":"api/Pipfile.lock","version":"1.0"}],"text":"textstring"}
+{"tag_name":"test1","references":[{"target":"api/Pipfile.lock","version":"1.0"},{"target":"api3/Pipfile.lock","version":"0.1"}]}
+{"tag_name":"test2"}
+{"tag_name":"test3","text":"text3"}
+{"tag_name":"alpha:alpha2:alpha3"}
