+++ api/app/tests/requests/upload_test/tag.txt
@@ -1,0 +1,2 @@
+{"tag_name":"teststring"}
+{"tag_name":"alpha"}
