+++ api/app/tests/requests/upload_test/tag_with_wrong_format.json
@@ -1,0 +1,1 @@
+{ "tag_name": "teststring", "tag_id": "test" }
