+++ api/app/tests/requests/upload_test/tag_with_wrong_format.jsonl
@@ -1,0 +1,3 @@
+{"tag_name":"teststring"}
+{"tag_name":"teststring1"},
+{"tag_name":"alpha"}
